Head reset 3 steps back + unused requirement removed

+++ dff/core/pipeline/service/group.py
@@ -45,7 +45,7 @@ class ServiceGroup(PipelineComponent):
         components: ServiceGroupBuilder,
         before_handler: Optional[ExtraHandlerBuilder] = None,
         after_handler: Optional[ExtraHandlerBuilder] = None,
-        timeout: Optional[int] = None,
+        timeout: Optional[float] = None,
         asynchronous: Optional[bool] = None,
         start_condition: Optional[StartConditionCheckerFunction] = None,
         name: Optional[str] = None,
